fix(speech): unblock hands-free CI
